♻️refactor: Naver 관련 DTO 패키지로 옮김 (#11)
