+++ src/main/java/kr/modernworld/modernworldv2/user/infrastructure/auth/naver/dto/NaverOpenAPIErrorDTO.java
@@ -1,4 +1,4 @@
-package kr.modernworld.modernworldv2.user.infrastructure.auth.naver;
+package kr.modernworld.modernworldv2.user.infrastructure.auth.naver.dto;
 
 // 아직은 안쓰임
 public record NaverOpenAPIErrorDTO(
