+++ src/main/java/kr/modernworld/modernworldv2/user/infrastructure/auth/naver/dto/NaverUserInfoResponseDTO.java
@@ -1,4 +1,4 @@
-package kr.modernworld.modernworldv2.user.infrastructure.auth.naver;
+package kr.modernworld.modernworldv2.user.infrastructure.auth.naver.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
